Added Douglas Fir nodes.
I need to entirely re-do the mapgen. There's so much chaos in that one file alone. I think that I will make this particular tree spawn higher up, in dry-soil biomes.

Or, I could scrap the concept of 'dry' and 'wet' soils and just make soil vary in composition. For example, coarse soil would hold fir and juniper, and maple would grow... in god-knows-what. It's a mysterious plant, indeed. Actually I might just replace maple with something else. If not that, then silty soil would hold it.

This is the longest commit comment I have ever written, and probably one of the most useless. Actually, this could be useful for future reference. Anyhow, goal is to get rid of present soil types, or to at least replace them.