Added more torches.
Now the douglasfir and juniper torches exist and can be crafted. I do not have any textures for them, yet. I will do that later on.

I would like to find a way to write my own version of walking_lights or illumination mods so that the player can light up nodes when holding torches or lamps.